requirements.txt
python-telegram-bot>=22.0
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
ollama>=0.1.7
pandas>=2.0.0
numpy>=1.24.0
